fix: share data from metrics across all dashboards
